% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppDoc.R, R/RcppExports.R
\name{CppFnc}
\alias{CppFnc}
\alias{g_char2int}
\alias{contiguous}
\alias{sum_g}
\alias{n_g}
\alias{max_g}
\alias{min_g}
\alias{dup_g}
\title{C++ functions (internal)}
\usage{
g_char2int(ch, start = 1L)

contiguous(g, error = FALSE)

sum_g(x, g, na_rm = FALSE, na_opt = FALSE, no_na = FALSE)

n_g(g)

max_g(x, g, na_rm = FALSE, na_opt = FALSE, no_na = FALSE)

min_g(x, g, na_rm = FALSE, na_opt = FALSE, no_na = FALSE)

dup_g(x, g, na_rm = FALSE)
}
\arguments{
\item{ch}{character (grouping) vector}

\item{start}{integer, starting value for integer grouping}

\item{g}{integer valued (grouping) vector}

\item{x}{numeric vector}

\item{na_rm}{boolean; remove missing values?}

\item{na_opt}{boolean; if na_rm is TRUE, should a grouping
value containing only missing x-values get summation NA (TRUE)
or else the standard value 0 (sum), -Inf (max), and Inf (min).}

\item{no_na}{boolean; is there a possibility of missing values?}
}
\value{
integer vector

a boolean

numeric vector of same length as x

numeric vector

numeric vector of same length as x

numeric vector of same length as x

boolean vector
}
\description{
More details here...
}
\details{
map character grouping variable into consecutive integer values, while also checking for contiguity

this function checks if x's unique values are contiguously arranged

sum numeric values by group

number of observations by group

calculate the maximum numeric value by group

calculate the minimum numeric value by group

check for duplicated values in x within values of g
}
