% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_calc.R
\name{grp_calc}
\alias{grp_calc}
\alias{grp_check}
\alias{grp_id}
\alias{grp_n}
\alias{grp_last}
\alias{grp_first}
\alias{grp_row}
\title{Calculations on grouping vector}
\usage{
grp_check(grp)

grp_id(grp)

grp_n(grp)

grp_last(grp)

grp_first(grp)

grp_row(grp)
}
\arguments{
\item{grp}{grouping vector}
}
\description{
Various functions to calculate things for a grouping vector. Important: grp
is assumed to be ordered such that equal values appear contiguously.
}
\details{
grp_check: check the grouping variable

grp_id: calculate an integer valued grouping vector

grp_n: calculate the number of rows per grouping value.

grp_last: calculate an indicator for last row per grouping value.

grp_first: calculate an indicator for first row per grouping value.

grp_row: calculate the row number per grouping value.
}
\examples{
id = sprintf("id\%s", c(2,2,1,4,4,4))
data.frame(
    id = id,
    grp_id = grp_id(id),
    grp_n = grp_n(id),
    grp_row = grp_row(id),
    grp_first = grp_first(id),
    grp_last = grp_last(id)
)
}
\seealso{
\code{\link{stats_by}}
}
