% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grp_calc.R
\name{shift}
\alias{shift}
\alias{lag1}
\alias{lead1}
\title{Shift vector}
\usage{
lag1(x, fill = NA)

lead1(x, fill = NA)
}
\arguments{
\item{x}{A vector.}

\item{fill}{Value to use for padding when the window goes beyond the input
length.}
}
\description{
Create shifted vector.
}
\details{
lag1 calcualtes lag by 1.

lead1 calculates lead by 1.
}
