% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_by.R
\name{stats_by}
\alias{stats_by}
\alias{sum_by}
\alias{max_by}
\alias{min_by}
\alias{duplicated_by}
\alias{n_by}
\alias{uniqueN_by}
\alias{anyTRUE_by}
\alias{lag1_by}
\alias{lead1_by}
\title{Fast statistics within grouping}
\usage{
sum_by(x, grp, na.rm = FALSE, NAopt = FALSE, noNA = FALSE, check = FALSE)

max_by(x, grp, na.rm = FALSE, NAopt = FALSE, noNA = FALSE, check = FALSE)

min_by(x, grp, na.rm = FALSE, NAopt = FALSE, noNA = FALSE, check = FALSE)

duplicated_by(x, grp, na.rm = FALSE, check = FALSE)

n_by(x, grp, na.rm = FALSE, check = FALSE)

uniqueN_by(x, grp, na.rm = FALSE, check = FALSE)

anyTRUE_by(x, grp, check = FALSE)

lag1_by(x, grp, check = FALSE)

lead1_by(x, grp, check = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{grp}{A grouping vector within values of which the statistic is to be
calculated. Should not contain missing values.}

\item{na.rm}{A logical value; remove missing values}

\item{NAopt}{A logical value; if missing values are removed, should the
result still be NA for the cases where x is completely missing?}

\item{noNA}{A logical value; in some cases x cannot contain missing values,
then this parameter can be set to TRUE to gain some speed.}

\item{check}{A logical value; check that grp is contiguous?}
}
\description{
\code{X_by} are functions to calculate the statistic X under a grouping
variable specified by grp. Important: x and grp are assumed to be ordered on
grp such that equal values appear consecutively, and there should be no
missing values in grp.
}
\details{
sum_by: summation.

max_by: maximum.

min_by: minimum.

duplicated_by: looks for duplicated values. If \code{na.rm = TRUE},
then missing values are never considered duplicates.

n_by: calculates number of rows. If \code{na.rm = TRUE} it will not
count rows where x is missing.

uniqueN_by: number of unique values.

anyTRUE_by: checks for any TRUE value. Note that NAs
will count as FALSE.

lag1_by: lag by 1.

lead1_by: lead by 1.
}
\seealso{
\code{\link{grp_calc}}
}
